refactor(input): extract keyboard handling to src/input.rs
- Move handle_input() and handle_dropped_files() to dedicated input.rs module
- Create impl Heike block for input handling methods in input.rs
- Make helper methods pub(crate) for cross-module accessibility
- Simplifies app.rs by removing 334 lines of keyboard logic
